refactor(workflow): fix directory naming and plugin feature location
Major structural improvements:

1. Directory Naming (src/jbom_new → src/jbom):
   - Package is still 'jbom', not 'jbom_new'
   - Updated all imports and references
   - Fixed environment.py, step definitions, and CLI

2. Plugin Features Organization:
   - Move POS feature to src/jbom/plugins/pos/features/
   - Plugins own their features, not top-level
   - Created behave.ini to discover plugin features
   - Updated BOOTSTRAP docs to show correct structure

3. Test Discovery:
   - behave.ini: lists core and plugin feature paths
   - pytest in pyproject.toml: src/jbom/plugins
   - Both configured to find plugin tests

4. Documentation Updates:
   - BOOTSTRAP.md: corrected directory structure
   - BOOTSTRAP-STEP2.md: plugin structure includes features/
   - Clarified where tests live and how to run them

Run tests:
- cd jbom-new && behave  (finds core + plugin features)
- cd jbom-new && pytest  (finds core + plugin unit tests)
