+++ jbom-new/behave.ini
@@ -1,0 +1,7 @@
+[behave]
+paths = features
+        src/jbom/plugins/pos/features
+default_format = pretty
+show_timings = true
+stdout_capture = false
+stderr_capture = false
