+++ jbom-new/src/jbom/plugins/pos/features/pos_generation.feature
